+++ R/preprocessing.R
@@ -590,6 +590,8 @@ GetResidual <- function(
 #'
 #' @inheritParams Read10X
 #' @inheritParams CreateSeuratObject
+#' @param data.dir Directory containing the H5 file specified by \code{filename}
+#' and the image data in a subdirectory called \code{spatial}
 #' @param filename Name of H5 file containing the feature barcode matrix
 #' @param slice Name for the stored image of the tissue slice
 #' @param filter.matrix Only keep spots that have been determined to be over
@@ -622,6 +624,10 @@ Load10X_Spatial <- function(
   to.upper = FALSE,
   ...
 ) {
+  if (length(x = data.dir) > 1) {
+    warning("'Load10X_Spatial' accepts only one 'data.dir'", immediate. = TRUE)
+    data.dir <- data.dir[1]
+  }
   data <- Read10X_h5(filename = file.path(data.dir, filename), ...)
   if (to.upper) {
     rownames(x = data) <- toupper(x = rownames(x = data))
@@ -1152,7 +1158,7 @@ Read10X_h5 <- function(filename, use.names = TRUE, unique.features = TRUE) {
   infile <- hdf5r::H5File$new(filename = filename, mode = 'r')
   genomes <- names(x = infile)
   output <- list()
-  if (!infile$attr_exists("PYTABLES_FORMAT_VERSION")) {
+  if (hdf5r::existsGroup(infile, 'matrix')) {
     # cellranger version 3
     if (use.names) {
       feature_slot <- 'features/name'
@@ -1425,9 +1431,10 @@ RunMoransI <- function(data, pos, verbose = TRUE) {
       error = function(x) c(1,1,1,1)
     )
   })
+  pcol <- ifelse(test = Rfast2.installed, yes = 2, no = 4)
   results <- data.frame(
     observed = unlist(x = results[1, ]),
-    p.value = unlist(x = results[2, ])
+    p.value = unlist(x = results[pcol, ])
   )
   rownames(x = results) <- rownames(x = data)
   return(results)