+++ R/visualization.R
@@ -203,7 +203,7 @@ DimHeatmap <- function(
 #' @importFrom stats median
 #' @importFrom scales hue_pal
 #' @importFrom ggplot2 annotation_raster coord_cartesian scale_color_manual
-#' ggplot_build aes_string
+#' ggplot_build aes_string geom_text
 #' @importFrom patchwork wrap_plots
 #' @export
 #'
@@ -657,7 +657,7 @@ ColorDimSplit <- function(
   CheckDots(..., fxns = 'DimPlot')
   tree <- Tool(object = object, slot = "BuildClusterTree")
   split <- tree$edge[which(x = tree$edge[, 1] == node), ][, 2]
-  all.children <- sort(x = tree$edge[, 2][! tree$edge[, 2] %in% tree$edge[, 1]])
+  all.children <- sort(x = tree$edge[, 2][!tree$edge[, 2] %in% tree$edge[, 1]])
   left.group <- DFT(tree = tree, node = split[1], only.children = TRUE)
   right.group <- DFT(tree = tree, node = split[2], only.children = TRUE)
   if (any(is.na(x = left.group))) {
@@ -739,12 +739,14 @@ ColorDimSplit <- function(
 #' @param ncol Number of columns for display when combining plots
 #' @param combine Combine plots into a single \code{\link[patchwork]{patchwork}ed}
 #' ggplot object. If \code{FALSE}, return a list of ggplot objects
+#' @param raster Convert points to raster format, default is \code{NULL} which
+#' automatically rasterizes if plotting more than 50,000 cells
 #'
 #' @return A \code{\link[patchwork]{patchwork}ed} ggplot object if
 #' \code{combine = TRUE}; otherwise, a list of ggplot objects
 #'
 #' @importFrom rlang !!
-#' @importFrom ggplot2 facet_wrap vars sym
+#' @importFrom ggplot2 facet_wrap vars sym labs
 #' @importFrom patchwork wrap_plots
 #'
 #' @export
@@ -783,7 +785,8 @@ DimPlot <- function(
   sizes.highlight = 1,
   na.value = 'grey50',
   ncol = NULL,
-  combine = TRUE
+  combine = TRUE,
+  raster = NULL
 ) {
   if (length(x = dims) != 2) {
     stop("'dims' must be a two-length vector")
@@ -800,7 +803,8 @@ DimPlot <- function(
   object[['ident']] <- Idents(object = object)
   orig.groups <- group.by
   group.by <- group.by %||% 'ident'
-  data[, group.by] <- object[[group.by]][cells, , drop = FALSE]
+  data <- cbind(data, object[[group.by]][cells, , drop = FALSE])
+  group.by <- colnames(x = data)[3:ncol(x = data)]
   for (group in group.by) {
     if (!is.factor(x = data[, group])) {
       data[, group] <- factor(x = data[, group])
@@ -827,7 +831,8 @@ DimPlot <- function(
         cells.highlight = cells.highlight,
         cols.highlight = cols.highlight,
         sizes.highlight = sizes.highlight,
-        na.value = na.value
+        na.value = na.value,
+        raster = raster
       )
       if (label) {
         plot <- LabelClusters(
@@ -851,7 +856,11 @@ DimPlot <- function(
             }
           )
       }
-      return(plot)
+      plot <- if (is.null(x = orig.groups)) {
+        plot + labs(title = NULL)
+      } else {
+        plot + CenterTitle()
+      }
     }
   )
   if (!is.null(x = split.by)) {
@@ -891,6 +900,12 @@ DimPlot <- function(
 #'  may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')
 #' @param split.by A factor in object metadata to split the feature plot by, pass 'ident'
 #'  to split by cell identity'; similar to the old \code{FeatureHeatmap}
+#' @param keep.scale How to handle the color scale across multiple plots. Options are:
+#' \itemize{
+#'   \item{"feature" (default; by row/feature scaling):}{ The plots for each individual feature are scaled to the maximum expression of the feature across the conditions provided to 'split.by'.}
+#'   \item{"all" (universal scaling):}{ The plots for all features and conditions are scaled to the maximum expression value for the feature with the highest overall expression.}
+#'   \item{NULL (no scaling):}{ Each individual plot is scaled to the maximum expression value of the feature in the condition provided to 'split.by'. Be aware setting NULL will result in color scales that are not comparable between plots.}
+#' }
 #' @param slot Which slot to pull expression data from?
 #' @param blend Scale and blend expression values to visualize coexpression of two features
 #' @param blend.threshold The color cutoff from weak signal to strong signal; ranges from 0 to 1.
@@ -942,6 +957,7 @@ FeaturePlot <- function(
   max.cutoff = NA,
   reduction = NULL,
   split.by = NULL,
+  keep.scale = "feature",
   shape.by = NULL,
   slot = 'data',
   blend = FALSE,
@@ -954,7 +970,8 @@ FeaturePlot <- function(
   by.col = TRUE,
   sort.cell = NULL,
   interactive = FALSE,
-  combine = TRUE
+  combine = TRUE,
+  raster = NULL
 ) {
   # TODO: deprecate fully on 3.2.0
   if (!is.null(x = sort.cell)) {
@@ -976,6 +993,10 @@ FeaturePlot <- function(
       reduction = reduction,
       slot = slot
     ))
+  }
+  # Check keep.scale param for valid entries
+  if (!(is.null(x = keep.scale)) && !(keep.scale %in% c("feature", "all"))) {
+    stop("`keep.scale` must be set to either `feature`, `all`, or NULL")
   }
   # Set a theme to remove right-hand Y axis lines
   # Also sets right-hand Y axis text label formatting
@@ -1211,12 +1232,14 @@ FeaturePlot <- function(
         pt.size = pt.size,
         cols = cols.use,
         shape.by = shape.by,
-        label = FALSE
+        label = FALSE,
+        raster = raster
       ) +
         scale_x_continuous(limits = xlims) +
         scale_y_continuous(limits = ylims) +
         theme_cowplot() +
-        theme(plot.title = element_text(hjust = 0.5))
+        CenterTitle()
+        # theme(plot.title = element_text(hjust = 0.5))
       # Add labels
       if (label) {
         plot <- LabelClusters(
@@ -1290,6 +1313,11 @@ FeaturePlot <- function(
             )
           )
         }
+      }
+      if (!(is.null(x = keep.scale)) && keep.scale == "feature" && !blend) {
+        max.feature.value <- max(data.single[, feature])
+        min.feature.value <- min(data.single[, feature])
+        plot <- suppressMessages(plot & scale_color_gradientn(colors = cols, limits = c(min.feature.value, max.feature.value)))
       }
       # Add coord_fixed
       if (coord.fixed) {
@@ -1420,6 +1448,11 @@ FeaturePlot <- function(
     }
     if (!is.null(x = legend) && legend == 'none') {
       plots <- plots & NoLegend()
+    }
+    if (!(is.null(x = keep.scale)) && keep.scale == "all" && !blend) {
+      max.feature.value <- max(data.plot[, features])
+      min.feature.value <- min(data.plot[, features])
+      plots <- suppressMessages(plots & scale_color_gradientn(colors = cols, limits = c(min.feature.value, max.feature.value)))
     }
   }
   return(plots)
@@ -1660,6 +1693,7 @@ IFeaturePlot <- function(object, feature, dims = c(1, 2), reduction = NULL, slot
 #' correlation between the two cells is displayed above the plot.
 #'
 #' @inheritParams FeatureScatter
+#' @inheritParams DimPlot
 #' @param cell1 Cell 1 name
 #' @param cell2 Cell 2 name
 #' @param features Features to plot (default, all features)
@@ -1681,7 +1715,8 @@ CellScatter <- function(
   highlight = NULL,
   cols = NULL,
   pt.size = 1,
-  smooth = FALSE
+  smooth = FALSE,
+  raster = NULL
 ) {
   features <- features %||% rownames(x = object)
   data <- FetchData(
@@ -1695,7 +1730,8 @@ CellScatter <- function(
     cols = cols,
     pt.size = pt.size,
     rows.highlight = highlight,
-    smooth = smooth
+    smooth = smooth,
+    raster = raster
   )
   return(plot)
 }
@@ -2813,6 +2849,12 @@ SpatialPlot <- function(
     stop("Please specific either group.by or features, not both.")
   }
   images <- images %||% Images(object = object, assay = DefaultAssay(object = object))
+  if (length(x = images) == 0) {
+    images <- Images(object = object)
+  }
+  if (length(x = images) < 1) {
+    stop("Could not find any spatial image information")
+  }
   if (is.null(x = features)) {
     if (interactive) {
       return(ISpatialDimPlot(
@@ -2891,12 +2933,6 @@ SpatialPlot <- function(
     )
     colnames(x = data) <- features
     rownames(x = data) <- Cells(x = object)
-  }
-  if (length(x = images) == 0) {
-    images <- Images(object = object)
-  }
-  if (length(x = images) < 1) {
-    stop("Could not find any spatial image information")
   }
   features <- colnames(x = data)
   colnames(x = data) <- features
@@ -3252,10 +3288,11 @@ BarcodeInflectionsPlot <- function(object) {
 #'
 #' @importFrom grDevices colorRampPalette
 #' @importFrom cowplot theme_cowplot
-#' @importFrom ggplot2 ggplot aes_string scale_size scale_radius geom_point
+#' @importFrom ggplot2 ggplot aes_string geom_point scale_size scale_radius
 #' theme element_blank labs scale_color_identity scale_color_distiller
 #' scale_color_gradient guides guide_legend guide_colorbar
 #' facet_grid unit
+#' @importFrom scattermore geom_scattermore
 #' @importFrom stats dist hclust
 #' @importFrom RColorBrewer brewer.pal.info
 #'
@@ -3316,6 +3353,7 @@ DotPlot <- function(
     names(x = feature.groups) <- features
   }
   cells <- unlist(x = CellsByIdentities(object = object, idents = idents))
+
   data.features <- FetchData(object = object, vars = features, cells = cells)
   data.features$id <- if (is.null(x = group.by)) {
     Idents(object = object)[cells, drop = TRUE]
@@ -4755,6 +4793,17 @@ SeuratTheme <- function() {
   return(DarkTheme() + NoLegend() + NoGrid() + SeuratAxes())
 }
 
+#' @importFrom ggplot2 theme element_text
+#'
+#' @rdname SeuratTheme
+#' @export
+#'
+#' @aliases CenterTitle
+#'
+CenterTitle <- function(...) {
+  return(theme(plot.title = element_text(hjust = 0.5), validate = TRUE, ...))
+}
+
 #' @inheritParams SeuratTheme
 #'
 #' @importFrom ggplot2 theme element_rect element_text element_line margin
@@ -5070,6 +5119,7 @@ WhiteBackground <- function(...) {
 # It happens to look good
 #
 # @param data A data frame being passed to ggplot2
+# @param raster If TRUE, point size is set to 1
 #
 # @return The "optimal" point size for visualizing these data
 #
@@ -5077,8 +5127,12 @@ WhiteBackground <- function(...) {
 # df <- data.frame(x = rnorm(n = 10000), y = runif(n = 10000))
 # AutoPointSize(data = df)
 #
-AutoPointSize <- function(data) {
-  return(min(1583 / nrow(x = data), 1))
+AutoPointSize <- function(data, raster = NULL) {
+  return(ifelse(
+    test = isTRUE(x = raster),
+    yes = 1,
+    no = min(1583 / nrow(x = data), 1)
+  ))
 }
 
 # Calculate bandwidth for use in ggplot2-based smooth scatter plots
@@ -5576,7 +5630,7 @@ FeaturePalettes <- list(
   'Seurat' = c('lightgrey', 'blue')
 )
 
-# Get colour aesththics from a plot for a certain geom
+# Get colour aesthetics from a plot for a certain geom
 #
 # @param plot A ggplot2 object
 # @param geom Geom class to filter to
@@ -5591,6 +5645,10 @@ GetColourAesthetics <- function(plot, geom = 'GeomPoint', plot.first = TRUE) {
       return(class(x = layer$geom)[1])
     }
   )
+  # handle case where raster is set to True
+  if (geom == "GeomPoint" && "GeomScattermore" %in% geoms) {
+    geom <- "GeomScattermore"
+  }
   geoms <- which(x = geoms == geom)
   if (!length(x = geoms)) {
     stop("Cannot find a geom of class ", geom)
@@ -5658,6 +5716,10 @@ GetXYAesthetics <- function(plot, geom = 'GeomPoint', plot.first = TRUE) {
       return(class(x = layer$geom)[1])
     }
   )
+  # handle case where raster is set to True
+  if (geom == "GeomPoint" && "GeomScattermore" %in% geoms){
+    geom <- "GeomScattermore"
+  }
   geoms <- which(x = geoms == geom)
   if (length(x = geoms) == 0) {
     stop("Cannot find a geom of class ", geom)
@@ -6610,7 +6672,7 @@ SetHighlight <- function(
   plot.order <- sort(x = unique(x = highlight), na.last = TRUE)
   plot.order[is.na(x = plot.order)] <- 'Unselected'
   highlight[is.na(x = highlight)] <- 'Unselected'
-  highlight <- as.factor(x = highlight)
+  highlight <- factor(x = highlight, levels = plot.order)
   return(list(
     plot.order = plot.order,
     highlight = highlight,
@@ -6675,14 +6737,19 @@ globalVariables(names = '..density..', package = 'Seurat')
 # @param smooth Make a smoothed scatter plot
 # @param rows.highight A vector of rows to highlight (like cells.highlight in SingleDimPlot)
 # @param legend.title Optional legend title
+# @param raster Convert points to raster format, default is \code{NULL}
+# which will automatically use raster if the number of points plotted is greater than
+# 50,000
+#
 # @param ... Extra parameters to MASS::kde2d
 #
 #' @importFrom stats cor
 # #' @importFrom MASS kde2d
 #' @importFrom cowplot theme_cowplot
 #' @importFrom RColorBrewer brewer.pal.info
-#' @importFrom ggplot2 ggplot geom_point aes_string labs scale_color_brewer
+#' @importFrom ggplot2 ggplot aes_string geom_point labs scale_color_brewer
 #' scale_color_manual guides stat_density2d aes scale_fill_continuous
+#' @importFrom scattermore geom_scattermore
 #'
 SingleCorPlot <- function(
   data,
@@ -6693,9 +6760,11 @@ SingleCorPlot <- function(
   rows.highlight = NULL,
   legend.title = NULL,
   na.value = 'grey50',
-  span = NULL
+  span = NULL,
+  raster = NULL
 ) {
-  pt.size <- pt.size <- pt.size %||% AutoPointSize(data = data)
+  pt.size <- pt.size %||% AutoPointSize(data = data, raster = raster)
+  raster <- raster %||% (nrow(x = data) > 50000)
   orig.names <- colnames(x = data)
   names.plot <- colnames(x = data) <- gsub(
     pattern = '-',
@@ -6717,6 +6786,7 @@ SingleCorPlot <- function(
     stop(msg, call. = FALSE)
   }
   plot.cor <- round(x = cor(x = data[, 1], y = data[, 2]), digits = 2)
+
   if (!is.null(x = rows.highlight)) {
     highlight.info <- SetHighlight(
       cells.highlight = rows.highlight,
@@ -6776,13 +6846,25 @@ SingleCorPlot <- function(
       guides(fill = FALSE)
   }
   if (!is.null(x = col.by)) {
-    plot <- plot + geom_point(
-      mapping = aes_string(color = 'colors'),
-      position = 'jitter',
-      size = pt.size
-    )
+    if (raster) {
+      plot <- plot + geom_scattermore(
+        mapping = aes_string(color = 'colors'),
+        position = 'jitter',
+        pointsize = pt.size
+      )
+    } else {
+      plot <- plot + geom_point(
+        mapping = aes_string(color = 'colors'),
+        position = 'jitter',
+        size = pt.size
+      )
+    }
   } else {
-    plot <- plot + geom_point(position = 'jitter', size = pt.size)
+    if (raster) {
+      plot <- plot + geom_scattermore(position = 'jitter', pointsize = pt.size)
+    } else {
+      plot <- plot + geom_point(position = 'jitter', size = pt.size)
+    }
   }
   if (!is.null(x = cols)) {
     cols.scale <- if (length(x = cols) == 1 && cols %in% rownames(x = brewer.pal.info)) {
@@ -6835,11 +6917,14 @@ SingleCorPlot <- function(
 # @param sizes.highlight Size of highlighted cells; will repeat to the length
 # groups in cells.highlight
 # @param na.value Color value for NA points when using custom scale.
+# @param raster Convert points to raster format, default is \code{NULL}
+# which will automatically use raster if the number of points plotted is greater than
+# 50,000
 #
 #' @importFrom cowplot theme_cowplot
 #' @importFrom RColorBrewer brewer.pal.info
-#' @importFrom ggplot2 ggplot aes_string labs geom_text guides
-#'  scale_color_brewer scale_color_manual element_rect guide_legend discrete_scale
+#' @importFrom ggplot2 ggplot aes_string geom_point labs guides scale_color_brewer
+#' scale_color_manual element_rect guide_legend discrete_scale
 #'
 SingleDimPlot <- function(
   data,
@@ -6856,9 +6941,11 @@ SingleDimPlot <- function(
   cells.highlight = NULL,
   cols.highlight = '#DE2D26',
   sizes.highlight = 1,
-  na.value = 'grey50'
+  na.value = 'grey50',
+  raster = NULL
 ) {
-  pt.size <- pt.size %||% AutoPointSize(data = data)
+  pt.size <- pt.size %||% AutoPointSize(data = data, raster = raster)
+  raster <- raster %||% (nrow(x = data) > 50000)
   if (length(x = dims) != 2) {
     stop("'dims' must be a two-length vector")
   }
@@ -6931,8 +7018,21 @@ SingleDimPlot <- function(
     )
     alpha.by <- NULL
   }
-  plot <- ggplot(data = data) +
-    geom_point(
+
+  plot <- ggplot(data = data)
+  plot <- if (isTRUE(x = raster)) {
+    plot + geom_scattermore(
+      mapping = aes_string(
+        x = dims[1],
+        y = dims[2],
+        color = paste0("`", col.by, "`"),
+        shape = shape.by,
+        alpha = alpha.by
+      ),
+      pointsize = pt.size
+    )
+  } else {
+    plot + geom_point(
       mapping = aes_string(
         x = dims[1],
         y = dims[2],
@@ -6941,9 +7041,12 @@ SingleDimPlot <- function(
         alpha = alpha.by
       ),
       size = pt.size
-    ) +
+    )
+  }
+  plot <- plot +
     guides(color = guide_legend(override.aes = list(size = 3))) +
-    labs(color = NULL)
+    labs(color = NULL, title = col.by) +
+    CenterTitle()
   if (label && !is.null(x = col.by)) {
     plot <- LabelClusters(
       plot = plot,